+++ src/metadata/definitions/interface-class-metadata.ts
@@ -3,4 +3,5 @@ import { TypeResolver } from "../../interfaces";
 
 export interface InterfaceClassMetadata extends ClassMetadata {
   resolveType?: TypeResolver<any, any>;
+  autoRegisteringDisabled: boolean;
 }